Add test cases for user model and implement coverage reporting
